refactor(sandbox): split large impl block and extract IpcChannel to submodule
Fixes #222

- Split the 167-line impl Sandbox block into 4 separate impl blocks
  (24 + 75 + 17 + 54 lines), all within the 100-line limit
- Extract spawn helpers: spawn_process(), wait_socket(), verify_engine()
  bringing spawn body from 64 to 11 lines (limit 50)
- Extract IpcChannel to src/permission/sandbox/ipc.rs submodule
- mod.rs reduced from 538 to 468 lines (limit 500)
